Combine Text tokens and preserve parse position
Preserving the parse positions means that it is possible to map the
output HTML to the original markdown document. This can be useful in
order to support parsing of sub-sections of the markdown, for example in
order to make live-editing possible. This also combines contiguous Text
tokens into a single Text token. This provides a small 1.3x speedup, and
a rather good 10x improvement in maximum memory usage (reported by
getMemMax) and a 100x improvement in the final tree size (reported by
getOccupiedMem) for parsing a largish 71k markdown document.